Added handler over NavigationController
Also added more debug info to the UI

+++ Core/Interfaces/IHeading.cs
@@ -2,5 +2,5 @@
 
 public interface IHeading
 {
-    double North { get; set; }
+    int North { get; set; }
 }
